+++ src/directives/formly-field.test.js
@@ -1189,6 +1189,20 @@ describe('formly-field', function() {
     });
   });
 
+  describe(`other things`, () => {
+    it(`should warn if you specify 'hide' in expressionProperties`, inject(($log) => {
+      scope.fields = [getNewField({expressionProperties: {hide: 'foo'}})];
+      compileAndDigest();
+      const log = $log.warn.logs[0];
+      expect($log.warn.logs).to.have.length(1);
+      expect(log[0]).to.equal('Formly Warning:');
+      expect(log[1]).to.equal(
+        'You have specified `hide` in `expressionProperties`. Use `hideExpression` instead'
+      );
+      expect(log[2]).to.equal(field);
+    }));
+  });
+
   function compileAndDigest(template = basicForm, context = scope) {
     el = $compile(template)(context);
     context.$digest();
@@ -1218,16 +1232,23 @@ describe('formly-field', function() {
       calledArgs = arguments;
     };
     test();
+    if (!calledArgs) {
+      throw new Error('Expected warning, but there was none');
+    }
     expect(calledArgs[0]).to.match(match);
     console.warn = originalWarn;
   }
 
   function shouldNotWarn(test) {
     var originalWarn = console.warn;
-    var callCount = 0;
-    console.warn = () => callCount++;
+    var calledArgs;
+    console.warn = function() {
+      calledArgs = arguments;
+    };
     test();
-    expect(callCount).to.equal(0);
+    if (calledArgs) {
+      throw new Error('Expected no warning, but there was one', calledArgs);
+    }
     console.warn = originalWarn;
   }
 });